alpha lat classifier stuff
