+++ dataAcq/startAudio.bat
@@ -8,4 +8,4 @@ if exist "%batdir%%execname%.exe" ( set buffexe="%batdir%%execname%.exe" )
 
 rem audio2ft deviceID host port
 rem N.B. config paramters from audio.cfg in this directory
-start /b "%execname%" %buffexe% 0 localhost 1972
+start /b "%execname%" %buffexe% 0 localhost 1972 %*
