+++ dataAcq/startJavaEventViewer.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 buffdir=`dirname $0`
-echo Starting: ${buffdir}/buffer/java/eventViewer.class $@
-java -cp ${buffdir}/buffer/java/BufferClient.jar:${buffdir}/buffer/java eventViewer $@
+echo Starting: ${buffdir}/buffer/java/EventViewer.class $@
+java -cp ${buffdir}/buffer/java/BufferClient.jar:${buffdir}/buffer/java EventViewer $@
