+++ dataAcq/startJavaSignalproxy.bat
@@ -1,4 +1,4 @@
 call ..\utilities\findJava.bat
 set batdir=%~dp0
 echo Starting: /buffer/java/SignalProxy.jar
-%javaexe% -cp "%batdir%buffer\java\BufferClient.jar;%batdir%buffer\java\SignalProxy.jar" nl.dcc.buffer_bci.SignalProxy localhost:1972
+%javaexe% -cp "%batdir%buffer\java\BufferClient.jar;%batdir%buffer\java\SignalProxy.jar" nl.dcc.buffer_bci.SignalProxy localhost:1972 %*
