+++ dataAcq/startPolybench.bat
@@ -1,1 +1,1 @@
-start "Polybench" "C:\Program Files\TMSi\TMSi FieldTrip\PolybenchRun.exe"
+start "Polybench" "C:\Program Files\TMSi\TMSi FieldTrip\PolybenchRun.exe" %*
