+++ dataAcq/startSignalProxy.bat
@@ -4,4 +4,4 @@ rem Search for the executable
 if exist "%batdir%buffer\bin\win32\csignalproxy.exe" ( set buffexe="%batdir%buffer\bin\win32\csignalproxy.exe" )
 if exist "%batdir%buffer\win32\csignalproxy.exe" ( set buffexe="%batdir%buffer\win32\csignalproxy.exe" )
 if exist "%batdir%csignalproxy.exe" ( set buffexe="csignalproxy.exe" )
-start /b "signalproxy" %buffexe%
+start /b "signalproxy" %buffexe% %*
