+++ java/sigProc/startAlphaLatContClassifier.sh
@@ -1,1 +1,1 @@
-java -cp "lib/BufferClient.jar:lib/commons-math3-3.4.jar:build/jar/AlphaLatContClassifier.jar" nl.dcc.buffer_bci.signalprocessing.AlphaLatContClassifier localhost:1972 res/clsfr_nf.txt 1280 100 
+java -cp "../../dataAcq/buffer/java/BufferClient.jar:lib/commons-math3-3.4.jar:build/jar/MatrixAlgebra.jar:build/jar/ContinuousClassifier.jar" nl.dcc.buffer_bci.signalprocessing.AlphaLatContClassifier localhost:1972 res/clsfr_nf.txt 1000 100
